+++ ASL422.h
@@ -4,6 +4,7 @@
 #include "ui_ASL232.h"
 #include <QTimer>
 
+
 class ASL422 : public QtResource
 {
     Q_OBJECT
@@ -18,7 +19,7 @@ public:
 
     int configRS422Format(short channelNum ,unsigned long baudrate,int wordlength,int parity,int stopBit);
     int sendRS422Data(short channelNum, short length ,unsigned char * pValue, unsigned int *pActualLength);
-    int recvRS422Data(short channelNum, short length, unsigned char * pValue, int *pActalLength,float *timeLab);
+    int recvRS422Data(short channelNum, short length, unsigned char * pValue, int *pActalLength,double *timeLab);
     int switch485VS422(int channel,RELAY_CONFIG eMode);
 private:
     int initRS422();
